Move new examples into examples/java
