+++ examples/java/src/main/java/org/apache/beam/examples/basic/ApproximateQuantilesExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import java.util.List;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
@@ -32,7 +34,7 @@ import org.slf4j.LoggerFactory;
 //   description: Demonstration of ApproximateQuantiles transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 44
+//   context_line: 46
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
@@ -40,32 +42,36 @@ import org.slf4j.LoggerFactory;
 //     - transforms
 //     - numbers
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        // Create a collection with numbers
-        PCollection<Integer> input = pipeline.apply(Create.of(1, 1, 2, 2, 3, 4, 4, 5, 5));
-        // This will produce a collection containing 5 values: the minimum value, Quartile 1 value, Quartile 2 value (median), Quartile 3 value, and the maximum value.
-        PCollection<List<Integer>> result = input.apply(ApproximateQuantiles.globally(5));
-        // [END main_section]
-        result.apply(ParDo.of(new LogOutput("PCollection numbers after ApproximateQuantiles.globally transform: ")));
-        pipeline.run();
+public class ApproximateQuantilesExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    // Create a collection with numbers
+    PCollection<Integer> input = pipeline.apply(Create.of(1, 1, 2, 2, 3, 4, 4, 5, 5));
+    // This will produce a collection containing 5 values: the minimum value,
+    // Quartile 1 value, Quartile 2 value (median), Quartile 3 value, and the
+    // maximum value.
+    PCollection<List<Integer>> result = input.apply(ApproximateQuantiles.globally(5));
+    // [END main_section]
+    result.apply(
+        ParDo.of(
+            new LogOutput("PCollection numbers after ApproximateQuantiles.globally transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
