+++ examples/java/src/main/java/org/apache/beam/examples/basic/CoGroupByKeyExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
@@ -31,7 +33,7 @@ import org.slf4j.LoggerFactory;
 //   description: Demonstration of CoGroupByKey transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 48
+//   context_line: 50
 //   categories:
 //     - Core Transforms
 //     - Joins
@@ -44,45 +46,43 @@ import org.slf4j.LoggerFactory;
 //     - pairs
 //     - group
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        // Create pairs
-        PCollection<KV<String, Integer>> pt1 = pipeline.apply(Create.of(
-            KV.of("a", 1),
-            KV.of("b", 2),
-            KV.of("b", 3),
-            KV.of("c", 4)));
-        PCollection<KV<String, String>> pt2 = pipeline.apply(Create.of(
-            KV.of("a", "apple"),
-            KV.of("a", "avocado"),
-            KV.of("b", "banana"),
-            KV.of("c", "cherry")));
+public class CoGroupByKeyExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    // Create pairs
+    PCollection<KV<String, Integer>> pt1 =
+        pipeline.apply(Create.of(KV.of("a", 1), KV.of("b", 2), KV.of("b", 3), KV.of("c", 4)));
+    PCollection<KV<String, String>> pt2 =
+        pipeline.apply(
+            Create.of(
+                KV.of("a", "apple"),
+                KV.of("a", "avocado"),
+                KV.of("b", "banana"),
+                KV.of("c", "cherry")));
 
-        final TupleTag<Integer> t1 = new TupleTag<>();
-        final TupleTag<String> t2 = new TupleTag<>();
-        PCollection<KV<String, CoGbkResult>> result =
-          KeyedPCollectionTuple.of(t1, pt1).and(t2, pt2)
-            .apply(CoGroupByKey.create());
-        // [END main_section]
-        result.apply(ParDo.of(new LogOutput("PCollection pairs after CoGroupByKey transform: ")));
-        pipeline.run();
+    final TupleTag<Integer> t1 = new TupleTag<>();
+    final TupleTag<String> t2 = new TupleTag<>();
+    PCollection<KV<String, CoGbkResult>> result =
+        KeyedPCollectionTuple.of(t1, pt1).and(t2, pt2).apply(CoGroupByKey.create());
+    // [END main_section]
+    result.apply(ParDo.of(new LogOutput("PCollection pairs after CoGroupByKey transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
