+++ examples/java/src/main/java/org/apache/beam/examples/basic/CountExample.java
@@ -15,18 +15,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.*;
-import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: Distinct
-//   description: Demonstration of Distinct transform usage.
+//   name: Count
+//   description: Demonstration of Count transform usage.
 //   multifile: false
 //   default_example: false
 //   context_line: 42
@@ -36,36 +37,32 @@ import org.slf4j.LoggerFactory;
 //   tags:
 //     - transforms
 //     - numbers
-//     - distinct
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
+public class CountExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    PCollection<Double> pc = pipeline.apply(Create.of(1.0, 2.0, 3.0, 4.0, 5.0));
+    PCollection<Long> count = pc.apply(Count.globally());
+    // [END main_section]
+    // Log values
+    count.apply(ParDo.of(new LogOutput("PCollection numbers after Count transform: ")));
+    pipeline.run();
+  }
 
-        // [START main_section]
-        // Create a collection with repeating numbers
-        PCollection<Integer> input = pipeline.apply(Create.of(1, 1, 2, 2, 3, 4, 4, 5, 5));
-        // Apply Distinct transform
-        PCollection<Integer> distinctNumbers = input.apply(Distinct.create());
-        // [END main_section]
-        // Log values
-        distinctNumbers.apply(ParDo.of(new LogOutput("PCollection numbers after Distinct.create transform: ")));
-        pipeline.run();
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
