+++ examples/java/src/main/java/org/apache/beam/examples/basic/CountPerKeyExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
@@ -25,45 +27,46 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: Combine
-//   description: Demonstration of Combine transform usage.
+//   name: CountPerKey
+//   description: Demonstration of Count.perKey transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 42
+//   context_line: 44
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
 //     - numbers
+//     - pairs
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        // Sum.ofIntegers() combines the elements in the input PCollection. The resulting PCollection, called sum,
-        // contains one value: the sum of all the elements in the input PCollection.
-        PCollection<Integer> pc = pipeline.apply(Create.of(1, 2, 3, 4, 5));
-        PCollection<Integer> sum = pc.apply(Combine.globally(Sum.ofIntegers()));
-        // [END main_section]
-        // Log values
-        sum.apply(ParDo.of(new LogOutput("PCollection numbers after Combine transform: ")));
-        pipeline.run();
+public class CountPerKeyExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    PCollection<KV<String, Integer>> input =
+        pipeline.apply(
+            Create.of(KV.of("a", 1), KV.of("a", 2), KV.of("b", 3), KV.of("b", 4), KV.of("b", 5)));
+    PCollection<KV<String, Long>> countPerKey = input.apply(Count.perKey());
+    // [END main_section]
+    // Log values
+    countPerKey.apply(ParDo.of(new LogOutput("PCollection numbers after Count transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
