+++ examples/java/src/main/java/org/apache/beam/examples/basic/FlatMapElementsExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import java.util.Arrays;
 import java.util.List;
 import org.apache.beam.sdk.Pipeline;
@@ -30,7 +32,7 @@ import org.slf4j.LoggerFactory;
 //   description: Demonstration of FlatMapElements transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 45
+//   context_line: 46
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
@@ -40,40 +42,40 @@ import org.slf4j.LoggerFactory;
 //     - strings
 //     - map
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        PCollection<String> lines = pipeline.apply(Create.of(
-            "the quick brown fox",
-            "jumps over the lazy",
-            "dog"
-        ));
-        // Use FlatMapElements to split the lines in PCollection<String> into individual words.
-        PCollection<String> words = lines.apply(FlatMapElements.via(
-            new InferableFunction<String, List<String>>() {
-            public List<String> apply(String line) throws Exception {
-                return Arrays.asList(line.split(" "));
-            }
-        }));
-        // [END main_section]
-        words.apply(ParDo.of(new LogOutput("PCollection elements after the transform: ")));
-        pipeline.run();
+public class FlatMapElementsExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    PCollection<String> lines =
+        pipeline.apply(Create.of("the quick brown fox", "jumps over the lazy", "dog"));
+    // Use FlatMapElements to split the lines in PCollection<String> into individual
+    // words.
+    PCollection<String> words =
+        lines.apply(
+            FlatMapElements.via(
+                new InferableFunction<String, List<String>>() {
+                  public List<String> apply(String line) throws Exception {
+                    return Arrays.asList(line.split(" "));
+                  }
+                }));
+    // [END main_section]
+    words.apply(ParDo.of(new LogOutput("PCollection elements after the transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
