+++ examples/java/src/main/java/org/apache/beam/examples/basic/GroupByKeyExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
@@ -25,43 +27,50 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: Mean
-//   description: Demonstration of Mean transform usage.
+//   name: GroupByKey
+//   description: Demonstration of GroupByKey transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 42
+//   context_line: 45
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - numbers
+//     - strings
+//     - pairs
+//     - group
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        PCollection<Double> pc = pipeline.apply(Create.of(1.0, 2.0, 3.0, 4.0, 5.0));
-        PCollection<Double> mean = pc.apply(Mean.globally());
-        // [END main_section]
-        // Log values
-        mean.apply(ParDo.of(new LogOutput("PCollection numbers after Mean transform: ")));
-        pipeline.run();
+public class GroupByKeyExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    PCollection<KV<String, String>> pt =
+        pipeline.apply(
+            Create.of(
+                KV.of("a", "apple"),
+                KV.of("a", "avocado"),
+                KV.of("b", "banana"),
+                KV.of("c", "cherry")));
+    PCollection<KV<String, Iterable<String>>> result = pt.apply(GroupByKey.create());
+    // [END main_section]
+    result.apply(ParDo.of(new LogOutput("PCollection pairs after GroupByKey transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
