+++ examples/java/src/main/java/org/apache/beam/examples/basic/KvSwapExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
@@ -29,7 +31,7 @@ import org.slf4j.LoggerFactory;
 //   description: Demonstration of KvSwap transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 41
+//   context_line: 43
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
@@ -37,34 +39,36 @@ import org.slf4j.LoggerFactory;
 //     - transforms
 //     - pairs
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        // Create key/value pairs
-        PCollection<KV<String, Integer>> pairs = pipeline.apply(Create.of(
-                KV.of("one", 1),
-                KV.of("two", 2),
-                KV.of("three", 3),
-                KV.of("four", 4)));
-        // Returns KV collection with keys and values swapped: PCollection<KV<K,V>> -> PCollection<KV<V,K>>
-        PCollection<KV<Integer,String>> swap = pairs.apply(KvSwap.create());
-        // [END main_section]
-        pairs.apply(ParDo.of(new LogOutput("PCollection element before KvSwap transform: ")));
-        swap.apply(ParDo.of(new LogOutput("PCollection element after KvSwap transform: ")));
-        pipeline.run();
+public class KvSwapExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    // Create key/value pairs
+    PCollection<KV<String, Integer>> pairs =
+        pipeline.apply(
+            Create.of(KV.of("one", 1), KV.of("two", 2), KV.of("three", 3), KV.of("four", 4)));
+    // Returns KV collection with keys and values swapped: PCollection<KV<K,V>> ->
+    // PCollection<KV<V,K>>
+    PCollection<KV<Integer, String>> swap = pairs.apply(KvSwap.create());
+    // [END main_section]
+    pairs.apply(ParDo.of(new LogOutput("PCollection element before KvSwap transform: ")));
+    swap.apply(ParDo.of(new LogOutput("PCollection element after KvSwap transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
