+++ examples/java/src/main/java/org/apache/beam/examples/basic/SampleExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
@@ -32,7 +34,7 @@ import org.slf4j.LoggerFactory;
 //   description: Demonstration of Sample transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 45
+//   context_line: 47
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
@@ -41,38 +43,42 @@ import org.slf4j.LoggerFactory;
 //     - pairs
 //     - group
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        // Create pairs
-        PCollection<KV<String, String>> pairs = pipeline.apply(Create.of(
+public class SampleExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    // Create pairs
+    PCollection<KV<String, String>> pairs =
+        pipeline.apply(
+            Create.of(
                 KV.of("fall", "apple"),
                 KV.of("spring", "strawberry"),
                 KV.of("winter", "orange"),
                 KV.of("summer", "peach"),
                 KV.of("spring", "cherry"),
                 KV.of("fall", "pear")));
-        // We use Sample.fixedSizePerKey() to get fixed-size random samples for each unique key in a PCollection of key-values.
-        PCollection<KV<String, Iterable<String>>> result = pairs.apply(Sample.fixedSizePerKey(2));
-        // [END main_section]
-        result.apply(ParDo.of(new LogOutput("PCollection pairs after Sample.fixedSizePerKey transform: ")));
-        pipeline.run();
+    // We use Sample.fixedSizePerKey() to get fixed-size random samples for each
+    // unique key in a PCollection of key-values.
+    PCollection<KV<String, Iterable<String>>> result = pairs.apply(Sample.fixedSizePerKey(2));
+    // [END main_section]
+    result.apply(
+        ParDo.of(new LogOutput("PCollection pairs after Sample.fixedSizePerKey transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
