+++ examples/java/src/main/java/org/apache/beam/examples/basic/ToStringExample.java
@@ -15,6 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.examples.basic;
+
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
@@ -25,51 +27,52 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 // beam-playground:
-//   name: GroupIntoBatches
-//   description: Demonstration of GroupIntoBatches transform usage.
+//   name: ToString
+//   description: Demonstration of ToString transform usage.
 //   multifile: false
 //   default_example: false
-//   context_line: 42
+//   context_line: 43
 //   categories:
 //     - Core Transforms
 //   complexity: BASIC
 //   tags:
 //     - transforms
-//     - strings
-//     - group
+//     - pairs
 
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        // Create pairs
-        PCollection<KV<String, String>> pairs = pipeline.apply(Create.of(
+public class ToStringExample {
+  public static void main(String[] args) {
+    PipelineOptions options = PipelineOptionsFactory.create();
+    Pipeline pipeline = Pipeline.create(options);
+    // [START main_section]
+    // Create key-value pairs
+    PCollection<KV<String, String>> pairs =
+        pipeline.apply(
+            Create.of(
                 KV.of("fall", "apple"),
                 KV.of("spring", "strawberry"),
                 KV.of("winter", "orange"),
                 KV.of("summer", "peach"),
                 KV.of("spring", "cherry"),
                 KV.of("fall", "pear")));
-        // Group pairs into batches of size 2
-        PCollection<KV<String, Iterable<String>>> result = pairs.apply(GroupIntoBatches.ofSize(2));
-        // [END main_section]
-        result.apply(ParDo.of(new LogOutput("PCollection pairs after GroupIntoBatches transform: ")));
-        pipeline.run();
+    // Use ToString on key-value pairs
+    PCollection<String> result = pairs.apply(ToString.kvs());
+    // [END main_section]
+    result.apply(ParDo.of(new LogOutput("PCollection key-value pairs after ToString transform: ")));
+    pipeline.run();
+  }
+
+  static class LogOutput<T> extends DoFn<T, T> {
+    private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
+    private final String prefix;
+
+    public LogOutput(String prefix) {
+      this.prefix = prefix;
     }
 
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
+    @ProcessElement
+    public void processElement(ProcessContext c) throws Exception {
+      System.out.println(prefix + c.element());
+      c.output(c.element());
     }
+  }
 }
