--- learning/beamdoc/java/WithKeys.java
@@ -1,73 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import org.apache.beam.sdk.Pipeline;
-import org.apache.beam.sdk.options.PipelineOptions;
-import org.apache.beam.sdk.options.PipelineOptionsFactory;
-import org.apache.beam.sdk.transforms.*;
-import org.apache.beam.sdk.values.KV;
-import org.apache.beam.sdk.values.PCollection;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-// beam-playground:
-//   name: WithKeys
-//   description: Demonstration of WithKeys transform usage.
-//   multifile: false
-//   default_example: false
-//   context_line: 42
-//   categories:
-//     - Core Transforms
-//   complexity: BASIC
-//   tags:
-//     - transforms
-//     - strings
-//     - pairs
-
-public class Task {
-    public static void main(String[] args) {
-        PipelineOptions options = PipelineOptionsFactory.create();
-        Pipeline pipeline = Pipeline.create(options);
-        // [START main_section]
-        PCollection<String> words = pipeline.apply(Create.of("Hello", "World", "Beam", "is", "fun"));
-        PCollection<KV<Integer, String>> lengthAndWord =
-            words.apply(WithKeys.of(new SerializableFunction<String, Integer>() {
-                @Override
-                public Integer apply(String s) {
-                    return s.length();
-                }
-            }));
-        // [END main_section]
-        lengthAndWord.apply(ParDo.of(new LogOutput("PCollection elements after WithKeys transform: ")));
-        pipeline.run();
-    }
-
-    static class LogOutput<T> extends DoFn<T, T> {
-        private static final Logger LOG = LoggerFactory.getLogger(LogOutput.class);
-        private final String prefix;
-
-        public LogOutput(String prefix) {
-            this.prefix = prefix;
-        }
-
-        @ProcessElement
-        public void processElement(ProcessContext c) throws Exception {
-            System.out.println(prefix + c.element());
-            c.output(c.element());
-        }
-    }
-}
